Add files via upload
In this project, I will focus on airline incidents. The data set for this assignment includes information on the cost of bird strikes. Used this data set to see if I can predict the cost of a bird strike (i.e., the Total Cost column in the data set) based on the attributes of the incident. This is important because this model can make a cost prediction as soon as a bird strike incident happens.